refactor: rename files (fixed spelling on library-error file), added scrollTo() for browser compat
