+++ projects/common/helpers/library-error.ts
