+++ projects/common/lib/components/date/date.component.ts
@@ -18,7 +18,7 @@ import startOfToday from 'date-fns/startOfToday';
 import addYears from 'date-fns/addYears';
 import subYears from 'date-fns/subYears';
 import subDays from 'date-fns/subDays';
-import { MoHCommonLibraryError } from '../../../helpers/library-errorr';
+import { MoHCommonLibraryError } from '../../../helpers/library-error';
 import { AbstractFormControl } from '../../models/abstract-form-control';
 import { compareAsc } from 'date-fns';
